Remove HTTP3 dependent code from the qpack implementation.
Motivation:

The QPACK decoder / encoder should not depend on HTTP3 and also throw their own exception type to make it easier to convert stuff to the right HTTP3 error code.

Modifications:

- Introduce QpackException and use it in the QPACK code
- Move Http3HeadersSink out of QpackDecoder and let it throw Http3Exception during header validation
- Adjust tests

Result:

QPACK not depends on http3 stuff anymore
