+++ src/main/java/io/netty/incubator/codec/http3/QpackDecoder.java
@@ -18,51 +18,20 @@ package io.netty.incubator.codec.http3;
 import io.netty.buffer.ByteBuf;
 import io.netty.util.AsciiString;
 
-import static io.netty.util.internal.ObjectUtil.checkPositive;
-import static io.netty.incubator.codec.http3.Http3Headers.PseudoHeaderName.getPseudoHeader;
-import static io.netty.incubator.codec.http3.Http3Headers.PseudoHeaderName.hasPseudoHeaderFormat;
-
 final class QpackDecoder {
 
-    private static final Http3Exception HEADER_ILLEGAL_INDEX_VALUE =
-        new Http3Exception("QPACK - illegal index value");
-    private static final Http3Exception NAME_ILLEGAL_INDEX_VALUE =
-        new Http3Exception("QPACK - illegal index value");
-
-    private static final long DEFAULT_MAX_HEADER_LIST_SIZE = 0xffffffffL;
+    private static final QpackException HEADER_ILLEGAL_INDEX_VALUE =
+            QpackException.newStatic(QpackDecoder.class, "getIndexedHeader(...)", "QPACK - illegal index value");
+    private static final QpackException NAME_ILLEGAL_INDEX_VALUE =
+            QpackException.newStatic(QpackDecoder.class, "getIndexedName(...)", "QPACK - illegal index value");
 
     private final QpackHuffmanDecoder huffmanDecoder = new QpackHuffmanDecoder();
 
-    private long maxHeaderListSize;
-
     /**
-     * Create a new instance.
-     */
-    QpackDecoder() {
-        this(DEFAULT_MAX_HEADER_LIST_SIZE);
-    }
-
-    /**
-     * Create a new instance.
-     */
-    QpackDecoder(long maxHeaderListSize) {
-        this.maxHeaderListSize = checkPositive(maxHeaderListSize, "maxHeaderListSize");
-    }
-
-    /**
-     * Decode the header block into header fields.
-     * <p>
+     * Decode the header block and add these to the {@link Sink}.
      * This method assumes the entire header block is contained in {@code in}.
      */
-    public void decodeHeaders(ByteBuf in, Http3Headers headers, boolean validateHeaders) throws Http3Exception {
-        Http3HeadersSink sink = new Http3HeadersSink(headers, maxHeaderListSize, validateHeaders);
-        decode(in, sink);
-
-        // Throws exception if detected any problem so far
-        sink.finish();
-    }
-
-    private void decode(ByteBuf in, Sink sink) throws Http3Exception {
+    public void decode(ByteBuf in, Sink sink) throws QpackException {
         // Required Insert Count
         // https://tools.ietf.org/html/draft-ietf-quic-qpack-19#section-4.5.1.1
         decodePrefixedInteger(in, 8);
@@ -86,7 +55,7 @@ final class QpackDecoder {
         }
     }
 
-    private long decodePrefixedInteger(ByteBuf in, int prefixLength) {
+    private static long decodePrefixedInteger(ByteBuf in, int prefixLength) {
         int nbits = (1 << prefixLength) - 1;
         int first = in.readByte() & nbits;
         if (first < nbits) {
@@ -105,7 +74,7 @@ final class QpackDecoder {
         return i;
     }
 
-    private void decodeIndexed(ByteBuf in, Sink sink) throws Http3Exception {
+    private static void decodeIndexed(ByteBuf in, Sink sink) throws QpackException {
         if ((in.getByte(in.readerIndex()) & 0x40) == 0x40) {
             final int staticIndex = (int) decodePrefixedInteger(in, 6);
             final QpackHeaderField field = getIndexedHeader(staticIndex);
@@ -115,7 +84,7 @@ final class QpackDecoder {
         }
     }
 
-    private void decodeLiteralWithNameReference(ByteBuf in, Sink sink) throws Http3Exception {
+    private void decodeLiteralWithNameReference(ByteBuf in, Sink sink) throws QpackException {
         if ((in.getByte(in.readerIndex()) & 0x10) == 0x10) {
             final int staticNameIndex = (int) decodePrefixedInteger(in, 4);
             final CharSequence name = getIndexedName(staticNameIndex);
@@ -126,23 +95,23 @@ final class QpackDecoder {
         }
     }
 
-    private void decodeLiteral(ByteBuf in, Sink sink) throws Http3Exception {
+    private void decodeLiteral(ByteBuf in, Sink sink) throws QpackException {
         final CharSequence name = decodePrefixedStringLiteral(in, (byte) 0x8, 3);
         final CharSequence value = decodePrefixedStringLiteral(in);
         sink.appendToHeaderList(name, value);
     }
 
-    private CharSequence decodePrefixedStringLiteral(ByteBuf in) throws Http3Exception {
+    private CharSequence decodePrefixedStringLiteral(ByteBuf in) throws QpackException {
         return decodePrefixedStringLiteral(in, (byte) 0x80, 7);
     }
 
-    private CharSequence decodePrefixedStringLiteral(ByteBuf in, byte mask, int prefix) throws Http3Exception {
+    private CharSequence decodePrefixedStringLiteral(ByteBuf in, byte mask, int prefix) throws QpackException {
         final boolean huffmanEncoded = (in.getByte(in.readerIndex()) & mask) == mask;
         final int length = (int) decodePrefixedInteger(in, prefix);
         return decodeStringLiteral(in, length, huffmanEncoded);
     }
 
-    private CharSequence decodeStringLiteral(ByteBuf in, int length, boolean huffmanEncoded) throws Http3Exception {
+    private CharSequence decodeStringLiteral(ByteBuf in, int length, boolean huffmanEncoded) throws QpackException {
         if (huffmanEncoded) {
             return huffmanDecoder.decode(in, length);
         }
@@ -151,7 +120,7 @@ final class QpackDecoder {
         return new AsciiString(buf, false);
     }
 
-    private CharSequence getIndexedName(int index) throws Http3Exception {
+    private static CharSequence getIndexedName(int index) throws QpackException {
         if (index <= QpackStaticTable.length) {
             final QpackHeaderField field = QpackStaticTable.getField(index);
             return field.name;
@@ -159,92 +128,14 @@ final class QpackDecoder {
         throw NAME_ILLEGAL_INDEX_VALUE;
     }
 
-    private QpackHeaderField getIndexedHeader(int index) throws Http3Exception {
+    private static QpackHeaderField getIndexedHeader(int index) throws QpackException {
         if (index <= QpackStaticTable.length) {
             return QpackStaticTable.getField(index);
         }
         throw HEADER_ILLEGAL_INDEX_VALUE;
     }
 
-    private static HeaderType validate(CharSequence name, HeaderType previousHeaderType) throws Http3Exception {
-        if (hasPseudoHeaderFormat(name)) {
-            if (previousHeaderType == HeaderType.REGULAR_HEADER) {
-                throw new Http3Exception(String.format("Pseudo-header field '%s' found after regular header.", name));
-            }
-
-            final Http3Headers.PseudoHeaderName pseudoHeader = getPseudoHeader(name);
-            if (pseudoHeader == null) {
-                throw new Http3Exception(String.format("Invalid HTTP/3 pseudo-header '%s' encountered.", name));
-            }
-
-            final HeaderType currentHeaderType = pseudoHeader.isRequestOnly() ?
-                    HeaderType.REQUEST_PSEUDO_HEADER : HeaderType.RESPONSE_PSEUDO_HEADER;
-            if (previousHeaderType != null && currentHeaderType != previousHeaderType) {
-                throw new Http3Exception(String.format("Mix of request and response pseudo-headers."));
-            }
-
-            return currentHeaderType;
-        }
-
-        return HeaderType.REGULAR_HEADER;
-    }
-
-    private enum HeaderType {
-        REGULAR_HEADER,
-        REQUEST_PSEUDO_HEADER,
-        RESPONSE_PSEUDO_HEADER
-    }
-
-    private interface Sink {
+    interface Sink {
         void appendToHeaderList(CharSequence name, CharSequence value);
-        void finish() throws Http3Exception;
-    }
-
-    private static final class Http3HeadersSink implements Sink {
-        private final Http3Headers headers;
-        private final long maxHeaderListSize;
-        private final boolean validate;
-        private long headersLength;
-        private boolean exceededMaxLength;
-        private Http3Exception validationException;
-        private HeaderType previousType;
-
-        Http3HeadersSink(Http3Headers headers, long maxHeaderListSize, boolean validate) {
-            this.headers = headers;
-            this.maxHeaderListSize = maxHeaderListSize;
-            this.validate = validate;
-        }
-
-        @Override
-        public void finish() throws Http3Exception {
-            if (exceededMaxLength) {
-                throw new Http3Exception(
-                    String.format("Header size exceeded max allowed size (%d)", maxHeaderListSize));
-            } else if (validationException != null) {
-                throw validationException;
-            }
-        }
-
-        @Override
-        public void appendToHeaderList(CharSequence name, CharSequence value) {
-            headersLength += QpackHeaderField.sizeOf(name, value);
-            exceededMaxLength |= headersLength > maxHeaderListSize;
-
-            if (exceededMaxLength || validationException != null) {
-                // We don't store the header since we've already failed validation requirements.
-                return;
-            }
-
-            if (validate) {
-                try {
-                    previousType = QpackDecoder.validate(name, previousType);
-                } catch (Http3Exception ex) {
-                    validationException = ex;
-                    return;
-                }
-            }
-
-            headers.add(name, value);
-        }
     }
 }
