+++ src/main/java/io/netty/incubator/codec/http3/QpackException.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2020 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package io.netty.incubator.codec.http3;
+
+import io.netty.util.internal.ThrowableUtil;
+
+/**
+ * Exception thrown if an error happens during QPACK processing.
+ */
+public final class QpackException extends Exception {
+
+    private QpackException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+
+    static QpackException newStatic(Class<?> clazz, String method, String message) {
+        return ThrowableUtil.unknownStackTrace(new QpackException(message, null, false, false), clazz, method);
+    }
+}
